refactor: move all MongoDB-specific connection logic into driver layer, add `createClient()` method to handle creating MongoClient
